pyproject.toml
[project]
name = "langgraph-qq-agent"
version = "0.1.0"
description = "LangGraph-based QQ Agent using NapCat + MCP"
requires-python = ">=3.11"

dependencies = [
    # LangChain 生态
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-google-genai>=2.0.0",

    # MCP 协议
    "mcp>=1.0.0",
    "langchain-mcp-adapters>=0.1.0",

    # 异步网络
    "websockets>=12.0",
    "httpx>=0.27.0",
    "aiofiles>=24.0",

    # 图片处理
    "Pillow>=10.0.0",

    # 语音处理
    "pysilk-mod>=1.6.0",

    # Web 框架 (沙盒)
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",

    # 配置管理
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",

    # 调试工具
    "langsmith>=0.1.0",

    # 日志
    "loguru>=0.7.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
